+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v1.2.0
+
+### Added
+
+- Worker option `--(not-)inject-stats-into-request` for injecting macroses into request object for accessing timestamp and allocated memory size (before request processing) values
+
 ## v1.1.1
 
 ### Added